Add a database and save woo objects
Instead of loading and discarding sessions and spots from the woo server each time, we now save them to the database. The spots are not reloaded with every request to /sessions anymore, the sessions however are. The next step is to regularly update the sessions in the database and not query the woo backend for sessions every time