Switch from Sinatra to Rails
This is mainly done in order to increase performance and making the app more mature. It makes way more sense to load the sessions with a scheduled cron job instead when a users actually visits a page. This would have been possible with Sinatra, however it would just have made things more complicated. The simple app that justified the choice for Sinatra over Rails is long gone, therefore the step towards Rails as a framework is taken